Exam description and basic functionality
Added a start screen for VA test that explains the test and start it
then the screen goes to a simple screen with a letter, textbox, and one button to end the test.
after "e" the letters are random press 1 for correct anything else for wrong.
Now i need to chance the sizes of the letters as the test progresses.